+++ posthog/api/test/test_organization_feature_flag.py
@@ -11,7 +11,7 @@ from posthog.models.cohort import Cohort
 from posthog.models.cohort.util import sort_cohorts_topologically
 from products.early_access_features.backend.models import EarlyAccessFeature
 from posthog.models.experiment import Experiment
-from posthog.models.feedback.survey import Survey
+from posthog.models.surveys.survey import Survey
 from posthog.models.team.team import Team
 from posthog.models.user import User
 from posthog.test.base import APIBaseTest, QueryMatchingTest, snapshot_postgres_queries
